Add new landing page Prepare people page from former landing page Create stub for motion page
