+++ src/styles/global.css
@@ -30,6 +30,9 @@
   --cl-theme-5: #222222;
   --cl-theme-unknown: #cccccc;
 
+  --cl-people-section: #fcbbdd;
+  --cl-motion-section: #e7eb75;
+
   --container-width: 920px;
 }
